Moving windows examples to their own folder, cleaning up exit validation failures

+++ examples/windows/vr_common.hpp
@@ -114,7 +114,7 @@ public:
             if (fpsTimer > 1000.0f) {
                 std::string windowTitle = getWindowTitle();
                 glfwSetWindowTitle(window, windowTitle.c_str());
-                lastFPS = frameCounter;
+                lastFPS = (float)frameCounter;
                 fpsTimer = 0.0f;
                 frameCounter = 0;
             }
